Add a rule hash to the key used to get metadata from cache
This commit reworks how we isolate component metadata rules between
projects. Instead of having a separate cache per project, which
turned out to be very expensive, we instead add a rule hash to
the key used for in-memory metadata caching. This hash is _only_
used for the in-memory cache and nothing else. It's a bit abusing
the infrastructure, but was made a necessary performance optimization.

In short, we keep a single in-memory cache, but when we read from the
cache, we're going to add a bit of salt corresponding to the rules
hash. If no project uses rules, it basically means we're going to reuse
the same instances. If not, it means we're going to read from the cache
a different instance, which is _later_ going to be processed by component
metadata rules.